Replace key codes that only exist on numeric keypads
Several of the examples let the user adjust parameters by means of the
GLFW_KEY_KP_ADD (keypad +) and GLFW_KEY_KP_SUBTRACT (keypad -) keys, but
these keys are not available on keyboards that lack a keypad, e.g. a
number of laptop keyboards and tenkeyless keyboards. As a result, users
of these keyboards can not adjust parameters in those examples.

In order to allow users of said keyboards to adjust the parameters on
those examples, this change modifies onKey() in the following manner:
- Parameter values are decreased by pressing the GLFW_KEY_MINUS key (-)
instead of GLFW_KEY_KP_SUBTRACT (keypad -)
- Parameter values are increased by pressing the GLFW_KEY_EQUAL key (=)
instead of GLFW_KEY_KP_ADD (keypad +). The rationale for this is that,
on US and UK keyboard layout, you hold SHIFT and strike = to type +.