Add ignore and attributes files and compile under MacOS
`glext.h` does not need to be included on MacOS, so it is only included
if the OS is not Apple.

The Cocoa, IOkit, CoreFoundation, and CoreVideo frameworks are needed
in order to compile and link glfw3. This has been added to the
`CMakeLists.txt` file if configuring on Apple.