Fix compilation with CodeBlocks/MinGW
MinGW does not come with OpenMP, so those examples are commented out
for the time being.

When using MinGW, the `Sleep()` function results in an unresolved
symbol when linking. These function calls are skipped when the
`__MINGW32__` symbol is defined.

Add instructions how to properly request C++11 and use of the
`M_PI` macro when using MinGW.