Minor edit and correct usage of sin() overloads
`sin()` by default operates on double in C. The original code casts the
result to `float`. This is slow and loses precision unnecessarily.

Rather than use `sinf()`, since the code is C++, it is better to cast
`currentTime` to `float` (using a `static_cast`) once and then use
std::sin(), which is overloaded for `float`.